git-svn-id: https://svn.terracotta.org/repo/tc/dso/branches/resource-management@20993 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ dso-common/src/main/java/com/tc/io/serializer/DSOSerializerPolicy.java
@@ -28,12 +28,12 @@ import java.util.Map;
 public final class DSOSerializerPolicy implements SerializerPolicy {
 
   private final Map                       class2SerializerDescriptor;
-  private final Map<Integer, Object>      id2Serializer;
+  private final Map<Byte, Object>         id2Serializer;
   private final DSOSerializerPolicy.SDesc objectSerializer;
 
   public DSOSerializerPolicy() {
     class2SerializerDescriptor = new HashMap();
-    id2Serializer = new HashMap<Integer, Object>();
+    id2Serializer = new HashMap<Byte, Object>();
 
     addSerializerMapping(Object.class, new ObjectSerializer());
     addSerializerMapping(ObjectID.class, new ObjectIDSerializer());
@@ -87,7 +87,7 @@ public final class DSOSerializerPolicy implements SerializerPolicy {
       this.serializer = serializer;
     }
 
-    int getID() {
+    byte getID() {
       return serializer.getSerializerID();
     }
   }
